added locking to theta buffer aggregator, factored out common locking code
